% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brocot.R
\name{example_sb_call}
\alias{example_sb_call}
\title{Compute Stern-Brocot Approximation for sqrt(2)}
\usage{
example_sb_call()
}
\value{
A data frame containing the rational approximation of the sqrt of 2,
along with details about the fraction, error, and Stern-Brocot path.
}
\description{
This function calls `coprimer::nearby_coprime()` to approximate the sqrt of 2
using the Stern-Brocot tree within a tolerance of 0.08.
}
\examples{
example_sb_call()
}
