% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot Histogram}
\usage{
plot_histogram(data, num_bins, col_name, title, x_label)
}
\arguments{
\item{data}{Data frame containing the column to plot.}

\item{col_name}{Column name as a string.}

\item{title}{Plot title.}

\item{x_label}{X-axis label.}
}
\description{
Generates a histogram for a given dataset.
}
