% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_histograms.R
\name{run_histograms}
\alias{run_histograms}
\title{Launch the Stern-Brocot Shiny App}
\usage{
run_histograms()
}
\description{
This function launches the Shiny app for visualizing Stern-Brocot approximations.
}
