% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_apps.R
\name{run_single_slit}
\alias{run_single_slit}
\title{Launch the Stern-Brocot Shiny App}
\usage{
run_single_slit()
}
\description{
This function launches the Shiny app for visualizing Stern-Brocot approximations.
}
