% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{single_slit_momentum_experiment}
\alias{single_slit_momentum_experiment}
\title{Launch the single slit momentum experiment}
\usage{
single_slit_momentum_experiment()
}
\description{
This function launches the Shiny app for visualizing momentum within a single slit.
}
