% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_apps.R
\name{single_slit_experiment}
\alias{single_slit_experiment}
\title{Launch the single slit position experiment}
\usage{
single_slit_experiment()
}
\description{
This function launches the Shiny app for visualizing position within a single slit.
}
